refactor(ui): update machine header to support tag forms
Separate the main machine/pool/tag view and header so that it can support tag header forms.

Fixes: canonical-web-and-design/app-tribe#741.
