+++ ui/src/app/base/components/node/MachinesHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MachinesHeader";
