+++ ui/src/app/machines/views/Machines.tsx
@@ -10,13 +10,7 @@ import NotFound from "app/base/views/NotFound";
 import type { MachineHeaderContent } from "app/machines/types";
 import machineURLs from "app/machines/urls";
 import MachineList from "app/machines/views/MachineList";
-import poolsURLs from "app/pools/urls";
-import PoolAdd from "app/pools/views/PoolAdd";
-import PoolEdit from "app/pools/views/PoolEdit";
-import Pools from "app/pools/views/Pools";
 import { FilterMachines } from "app/store/machine/utils";
-import tagURLs from "app/tags/urls";
-import Tags from "app/tags/views/Tags";
 
 const Machines = (): JSX.Element => {
   const history = useHistory();
@@ -29,7 +23,6 @@ const Machines = (): JSX.Element => {
   const [headerContent, setHeaderContent] =
     useState<MachineHeaderContent | null>(null);
   const actionSelected = headerContent?.view[0] === "machineActionForm";
-  const previousPath = usePrevious(location.pathname);
   const previousActionSelected = usePrevious(actionSelected);
 
   const setSearchFilter = useCallback(
@@ -40,13 +33,6 @@ const Machines = (): JSX.Element => {
     },
     [history, setFilter]
   );
-
-  useEffect(() => {
-    // When the page changes (e.g. /pools -> /machines) then update the filters.
-    if (location.pathname !== previousPath) {
-      setFilter(FilterMachines.filtersToString(currentFilters));
-    }
-  }, [location.pathname, currentFilters, previousPath]);
 
   useEffect(() => {
     if (actionSelected !== previousActionSelected) {
@@ -83,19 +69,6 @@ const Machines = (): JSX.Element => {
               setSearchFilter={setSearchFilter}
             />
           )}
-        />
-        <Route exact path={poolsURLs.pools} render={() => <Pools />} />
-        <Route exact path={poolsURLs.add} render={() => <PoolAdd />} />
-        <Route
-          exact
-          path={poolsURLs.edit(null, true)}
-          render={() => <PoolEdit />}
-        />
-        <Route exact path={tagURLs.tags.index} render={() => <Tags />} />
-        <Route
-          exact
-          path={tagURLs.tag.index(null, true)}
-          render={() => <Tags />}
         />
         <Route path="*" render={() => <NotFound />} />
       </Switch>
