+++ ui/src/app/pools/components/TagHeaderForms/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TagHeaderForms";
