+++ ui/src/app/pools/views/PoolEdit/PoolEdit.tsx
@@ -1,10 +1,9 @@
 import { Spinner } from "@canonical/react-components";
 import { useSelector } from "react-redux";
 
-import PoolForm from "../PoolForm";
-
 import ModelNotFound from "app/base/components/ModelNotFound";
 import { useGetURLId } from "app/base/hooks/urls";
+import PoolForm from "app/pools/components/PoolForm";
 import poolURLs from "app/pools/urls";
 import poolSelectors from "app/store/resourcepool/selectors";
 import { ResourcePoolMeta } from "app/store/resourcepool/types";
