+++ ui/src/app/pools/views/PoolList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PoolList";
