+++ ui/src/app/tags/components/TagsHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TagsHeader";
