+++ ui/src/app/tags/constants.ts
@@ -1,0 +1,3 @@
+export const TagHeaderViews = {
+  AddTag: ["addTagForm", "addTag"],
+} as const;
